requirements.txt
aws-cdk-lib==2.137.0
constructs>=10.0.0,<11.0.0
opensearch-py==2.5.0
requests-aws4auth==1.2.3
utils==1.0.2
elasticsearch==7.13.4